booktown/requirements.txt
asgiref==3.6.0
Django==4.2.1
django-cors-headers==3.14.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.4
